webadmin/userportal: VM compatibility version UI
This patch is part of the series introducing VM compatibility version
feature. This feature allows to customize a VM to its own compatibility
version unrelated to one of the cluster. VM compatibility version is
limited by Data Center and capabilities of the host the VM will be able
to run on.

This patch adds UI for editing/viewing the VM compatibility version.

Change-Id: I0d2d1b19898ede60656d78d23d5c95b113a71cc6
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>