core: Add VM compatibility version to ChangeVmClusterValidator
This patch is part of the series introducing VM compatibility version
feature. This feature allows to customize a VM to its own compatibility
version unrelated to one of the cluster. VM compatibility version is
limited by Data Center and capabilities of the host the VM will be able
to run on.

This patch adds VM compatibility version support to
ChangeVmClusterValidator.

Change-Id: Ic5e35963ba47aefd17a69495ed4f4187abbc34ad
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>