engine,webadmin: do not invoke HostSetupNetworksCommand where no changes were made
• discard from HostSetupNetworksCommandParameters all bonds and
network attachments, which are sent to be updated, but they're
equal to already existing Bonds resp. NetworkAttachment in the system.
• added method to HostSetupNetworksCommandParameters to figure out if
there's any change to be done in them, and this method is  used both on UI and
bll to figure out, whether changes needs to be done to the system.

Change-Id: Icc4c17171ed37cea79ea80aa65b7f2a056ca407c
Bug-Url: https://bugzilla.redhat.com/1254320
Signed-off-by: Martin Mucha <mmucha@redhat.com>