frontend: default volume format selection in new template window
When making a template from a vm via the ui, one can choose the volume
format of each template disk (92665e0af829d25273d5d45669847569eae6f5cc).
Before this feature, the format of each disk was the same as its base
snapshot, but now it's set to QCOW2 by default.
This patch restores the old behavior such that the format is chosen to
be the same as the base snapshot's format by default.

Change-Id: I364a8f23f75b583c1f5b24051d374f32fa53a39d
Bug-Url: https://bugzilla.redhat.com/1257240
Signed-off-by: Idan Shaby <ishaby@redhat.com>