core: Implement cold reboot support
Cold Reboot = Graceful Shutdown + Fresh Run

When VM's configuration has changed (next run configuration exists or
a run-once case) we set a flag in VmManager, which is detected in
monitoring when VM exists gracefully and control is passed to
ColdRebootAutoStartVmsRunner.

The periodic restart-attempt logic from AutoStartVmsRunner class was split into
abstract base class and re-used it in the new ColdRebootAutoStartmsRunner.

Change-Id: Ie2f7289fe7a54d5df43ccad1d6c0206308830661
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=751854
Signed-off-by: Martin Betak <mbetak@redhat.com>