restapi: Constraint parser
This patch improves adds a parser for constraint expressions, so that
constraints can be added to the model using Java code in default
methods. For example the constraint that says that when adding a virtual
machine the name is mandatory can be now expressed adding a method like
this to the "Add" method of the "VmService":

  interface Add {
    /**
     * Creates a new VM ...
     */
    @In @Out Vm vm();

    /**
     * This defines a new constraint that says that the input
     * of the `Add` operation should contain a name.
     */
    @Required @In
    default void nameIsMandatory() {
      assert vm().name() != null;
    }
  }

Later patches will move the specification of constraints from the
"rsdl_metadata.yaml" file to the metamodel.

Change-Id: Iea6d2e535f89a7c166eb17e0f951bf06699ccd8e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>