core: Support recursive remove of cinder volumes.
Implementing a recursive way to remove Cinder volumes.
As part of this implementation the behviour should be as follow:
When trying to remove a Cinder disk we aggragate all the volumes of the
Cinder disk, snapshots and Cinder volumes (commited snapshots).
Then we execute them one at a time, if there is a failure in the middle
we change the disk status to illegal and fail the command.

Once an illegal Cinder disk is being removed we only remove the Cinder
disk and its volumes from the DB, without removing it from the Cinder
provider.

Change-Id: Icc5c19c0fa06f2fb40423072e609b3c0ca7f28f3
Bug-Url: https://bugzilla.redhat.com/1278126
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>