restapi: Import VM -  support volume type/format
Adding support for changing the allocation policy of disks imported from
an export domain as part of importing a VM.

The caller can specify for each disks the format and allocation policy.
Note that "collapse_snapshots" should be true.

For example, to force one of the disks to be preallocated:
POST
/storagedomains/{storagedomain:id}/vms/{vm:id}/import;collapse_snapshots=true

  <action>
   <storage_domain>
    <name>mydata</name>
   </storage_domain>
   <cluster>
    <name>mycluster</name>
   </cluster>
   <vm>
    <disks>
     <disk id="792f8ca8-3932-429b-9660-006ba144bc72">
      <format>raw</format>
      <sparse>false</sparse>
     </disk>
    </disks>
   </vm>
  </action>

Change-Id: I0cd0dc95bfb938d66e8e1ba1ad56116e4dc173f3
Bug-Url: https://bugzilla.redhat.com/998607
Signed-off-by: Fred Rolland <frolland@redhat.com>