engine: make gateway as sync property
The gateway will be reported as out of sync in
case of a difference between reported and defined
gateway.  please note that the gateway's syncness
is being checked only in case both defined and
reported boot protocol are static.  The reason is
that in case that boot protocol is DHCP then the
gateway is being determined locally, contrary when
it is static which is being defined by the user.

Change-Id: I2a8b206692e459992630d7b8558801e9d6e0b426
Bug-Url: https://bugzilla.redhat.com/1270688
Signed-off-by: Eliraz Levi <elevi@redhat.com>