core: detaching master domain cleanup
The LinqUtils call in
StorageDomainCommandBase#isDetachAllowed(boolean) is useless - it
retrieves a domain from the pool that has the same ID of the domain
being removed (i.e., the exact same domain) and an "Active" status.
The previous call to checkStorageDomainStatus in canDetachDomain
ensures its status will be "Maintenance" or "Inactive", so it's known
to return an null.

Since its result is known, it can be skipped, and save a relatively
expensive database call.

Change-Id: Iff87f89170d555d98ec071b42e46f3dff1883b1b
Signed-off-by: Allon Mureinik <amureini@redhat.com>