restapi: Add stream methods to the model
The model classes have methods that return lists. For example, the
"StructType" class has a "getAttributes" method that returns its list of
attributes. To avoid issues these lists are copies of the ones used
internally. This patch changes the model classes so that they also have
methods that returns streams, as those are nicer in Java 8 and no copy
of the elements of the lists is required.

Change-Id: I975cd02b9771665fb07c8c161bf814ea2dbe9258
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>