core: Remove GetVGList*
GetVGList query returns all the VGs a host can access that are not
correlated to storage domains saved in the database. It is only used in
BackendHostStorageResource, where it is used to retrieve connection
information. This information is redundant, as it's always used together
with the connections related to GetDeviceList, which retrieves devices
related to VGs too, anyway.

This patch removes this entire redundant query, cleaning up the code and
improving the performance of hosts/{hostId}/storage and
hosts/{hostId}/storage/{storageId} by a significant factor.

This patch includes:
1. Removing the redundant calls to GetVgList in
   BackendHostStorageResource and amending the test accordingly.
2. Removing GetVgListQuery class and the corresponding constant from
   VdcQueryType, now that there are no more usages for it.
3. Removing GetVGListVDSCommand, the corresponding constant in
   VDSCommandType, the corresponding methods in IVdsServer and its
   implementations and the related VGListReturnForXmlRpc class, now that
   there are no more usages for it.

Change-Id: Ie16e0da825e74defd0aada439b58bb25e93bc802
Signed-off-by: Allon Mureinik <amureini@redhat.com>