core: fix remove vm from export domain
The injection of BackendInternal broke import from export domain.
The problem was that in the constructor of RemoveVmCommand we called
getVm() that invokes a query to get the VM from the export domain.
However, BackendInternal is not injected at this stage so NPE is
thrown. The solution is to extract the initialization code in
RemoveVmCommand into the 'init' method that is called after the
object is constructed, including its injected fields.

Change-Id: Ieb4cd5f4b301d73a996ad4acc505fd8bb42e29fd
Signed-off-by: Arik Hadas <ahadas@redhat.com>