core: MoveOrCopyDisk - attempt to perform rollback for vm/floating disk
Currently when executing MoveOrCopyDisk for a non template disk no rollback is
done and the disk will remain in the engine as a legit disk, while for
template disk the mapping of the disk on the other domain will be
deleted from the db.

This patch adds an attempt to perform a rollback when copying a non
template disk, the rollback when copying a template disk requires some
other modifications and will be handled on a different patch.

Change-Id: Id6727f513efca391ff843a1665dabb7826533706
Bug-Url: https://bugzilla.redhat.com/1232396
Signed-off-by: Liron Aravot <laravot@redhat.com>