restapi: Add GET and HEAD methods to the system resource
The "SystemResource" interface generated by the JAX-RS code generator
doesn't include the following two methods:

  @GET
  Response get();

  @HEAD
  Response head();

Instead of that the methods and the annotations are part of the
implementation class. In previous versions of Resteasy this worked
correctly, but in the newer version included in WildFly the annotations
aren't recognized. As a result requests to /ovirt-engine/api return
nothing.

To solve this issue this patch changes the JAX-RS code generator so that
it explicitly adds these two methods to the "SystemResource" interface.

Change-Id: I7a0dfa3bb7e3d7a3087580e79d2fe576a2daec34
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>