engine: update setupNetworks heartbeat policy only for mgmt networks
Updating the heartbeat policy during the HostSetupNetworks command
only makes sense when the connection between the engine and host is
modified. This should only be the case when the management network
is modified.
This patch changes the rpc broker to only reset the heartbeat policy when
the management network is changed in any way.

Change-Id: I49ef2e8d520c5cfff769375a0d5f9428179e87d8
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1262051