core: VM compatibility version
This patch is part of the series introducing VM compatibility version
feature. This feature allows to customize a VM to its own compatibility
version unrelated to one of the cluster. VM compatibility version is
limited by Data Center and capabilities of the host the VM will be able
to run on.

In this patch:

1. custom_compatibility_version field is added to VM-related DB tables,
views and Java business entities.

2. vdsGroupCompatibilityVersion field was added to VmTemplate to
simplify implementation of the methods described below.

3. VM.getCompatibilityVersion() method is introduced. It returns
the VM compatibility version, if set, otherwise - the cluster one.
The similar method is added to VmTemplate. These methods should be
used in every place where cluster compatibility version is checked
currently, with exception to a functionality that must depend only on
cluster compatibility version (such as networking or storage).

Actual usage of these methods requires a lot of simple changes
throughout the code. These changes were split off into the separate
patch.

4. Some commands (such as add/update/import VM) work with VMs that were
not commited into the database yet or their cluster may be changed. The
method described above cannot be used in this case, so the effective
compatibility version for these VMs is calculated separately and cached
in the command.

Change-Id: I7fd2f68f1546f5953457da38e4df1e3c4b991de4
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>