webadmin: edit pool dialog was not always filled properly
The HW part should be filled by InstanceTypeManager and the SW part by behavior
classes. In the current implementation the instance type manager was bypassed
in some cases and only used as a signal when to pass the control back to it.

Fixed by moving the HW related logic to Instance TypeManager where it belongs
to and let the SW part in the behavior.

Change-Id: I0866a3d23ba946c5a9f497b03a9b4537aea750ee
Bug-Url: https://bugzilla.redhat.com/1282853
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>